WebRTC stats should be in milliseconds https://bugs.webkit.org/show_bug.cgi?id=172644
Patch by Youenn Fablet <youenn@apple.com> on 2017-05-26
Reviewed by Eric Carlson.

Source/WebCore:

Covered by updated tests.

* Modules/mediastream/RTCStatsReport.h:
* Modules/mediastream/RTCStatsReport.idl:
* Modules/mediastream/libwebrtc/LibWebRTCMediaEndpoint.cpp:
(WebCore::fillRTCStats):

LayoutTests:

* TestExpectations:
* webrtc/video-stats.html: Making it less flaky prone.

Canonical link: https://commits.webkit.org/189571@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@217519 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/Modules/mediastream/RTCStatsReport.h
@@ -53,7 +53,7 @@ public:
         Certificate
     };
     struct Stats {
-        unsigned long long timestamp;
+        double timestamp;
         Type type;
         String id;
     };
